02_meme_gen_starter/requirements.txt
Flask==1.1.1
numpy==1.17.4
pandas==0.25.3
pdftotext==2.1.2
pdfminer==20191103
Pillow==6.2.1
python-docx==0.8.10
requests==2.22.0

